06/22/2019 - Update For Github
I have went through Gamehub and cleaned up the project the best that I
could. I look back on the programming that was completed the years
before and I see all of the errors and organization optimizations that
need to be made.

This will be the start of upgrading Gamehub and having it fully
functional as a true Gamehub.